+++ src/R-yyjson-serialize.h
@@ -78,6 +78,7 @@ typedef struct {
   unsigned int num_specials;
   unsigned int yyjson_write_flag;
   bool fast_numerics;
+  bool json_verbatim;
 } serialize_options;
 
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~