+++ NEWS.md
@@ -1,6 +1,6 @@
 
 
-# yyjsonr 0.1.20.9008 2025-04-26
+# yyjsonr 0.1.20.9009 2025-04-26
 
 * [9000] Get size of gzipped file in an endian-neutral way. Issue #39
 * [9001] Fix underflow issue when trying to verbosely report an error at position 0
@@ -15,6 +15,8 @@
 * [9007] Properly support `YYJSON_READ_NUMBER_AS_RAW` to read all numbers as strings
 * [9008] Add option `digits_promote` to control decimal places when numbers in
          JSON []-arrays are promoted to strings.
+* [9009] Add optino `json_verbatim` to write strings (with class "json") directly
+         into the output
 
 # yyjsonr 0.1.20 2024-04-10